[fix][pulsar-broker] Fix RawReader hasMessageAvailable returns true when no messages ---
*Motivation*

The RawReader hasMessageAvailable will return true when all the
messages have been consumed. And that will cause the readNextAsync
blocked and the process never recovered.
In the ConsumerImpl, we update the lastDequeuedMessageId in the
messageProcess. The messageReceived method rewrites by the RawReader,
we should update the lastDequeuedMessageId in the RawReader as well.

(cherry picked from commit ffd1a3cb40216646c747f96f1fedf3cb47c5e537)