[Fix] Allow the content be empty when deserialize in MetadataCache ---
### Motivation

When using admin to create topic, it will set the content to byte[0]
then put it into the value. Sometimes we need to update the value
then we will received the deserailization error.
So we need to check the value type to make it as optional empty
when the value is null.